vscode: build: add option to not use vivado
Use an environment variable (that should be defined in the tasks.json
file) to make it possible to not look for a vivado installation. The
motivation for this was that compiling a multiplatform ARM defconfig was
failing due to some issue in the Vivado toolchain (might just be an
issue with the installation on my system).

Signed-off-by: Nuno <noname.nuno@gmail.com>