git-publish: add script
This script is wrapper around git-publish [1]. This makes it much easier
to send patchseries to linux (for example) using git send-mail. It will
save cover-letters, keeps track of the series version being sent and
also stores CCs and TOs. To get the maintainers and mailing lists to whom
the patches must be sent I based myself in [2].

[1]: https://github.com/stefanha/git-publish
[2]: https://github.com/andy-shev/home-bin-tools/blob/master/ge2maintainer.sh
Signed-off-by: Nuno Sá <noname.nuno@gmail.com>