tmux.conf: fix double and triple click mouse buttons
After doing a mouse double or triple click to select some word, the next
mousedrag event would fail. I'm not really sure why that was happing
but it seems somehow related with the "next-space-end" command. Anyways,
we can actually redo the bindings in order to fix this behavior and also
improve the word selection. We can do it by changing the
'word-separators' from the default value to ' ,;' which is perfectly
acceptable. With this, we can now use commands as 'previous-word' and
'select-word' to make it reliably work.

Note that with this change, we can now click anywhere in the word to
fully select it (before we had to do it on it's first character).

Signed-off-by: Nuno Sá <noname.nuno@gmail.com>