git_publish: add --end-series command
Add a command to end a git series. Essentially, given a branch, this
will look for all the tags (series versions) and deletes them. The tags
are named as ${branch}-v${version_number}. If no branch is given, we
default to the one currently pointed to by HEAD. The branch will also be
deleted...

Note that this is given as a typical option even though it feels like a
command. If we start to add more commands, this can be re-thought so it
does not look like a regular top level option.

Signed-off-by: Nuno Sá <noname.nuno@gmail.com>