git-series-push: fix base branch not being set
As the script is running with -e, if we cannot find any base branch
in the branch configuration, we'll get an error and exit before
configuring the base branch. And this was all silent as the error comes
from a subshell invocation.

To prevent that, let's never error out (using || true) when looking for
the base branch. We will still have an empty string if no branch is set
which allows us to do the proper checks and act accordingly.

Signed-off-by: Nuno Sa <noname.nuno@gmail.com>