git-series-push: remove get_maintainer.pl options
These options were never really used so let's just remove that
possibility for now. It will make a future change (of only calling the
script on version 1 of the series) easier.

We can always re-add this option again but let's do it only if there's a
real need for it.

While at it, rephrased the help message to remove mentions to
get_maintainers and to state that this tool can be used in more things
than linux.

Signed-off-by: Nuno Sa <noname.nuno@gmail.com>