remote-connect: add new script
This script is used for remote connection. It's intended for connect to
my work laptop. As in the office, the IP constantly changes, it becomes
annoying to have to manually change the desktop file all the time. Hence
the script will first inspect if the saved IP matches the configured mac
to access (using arp). If that does not work, we use 'nmap' to get the
new ip and connect.

Signed-off-by: Nuno Sa <noname.nuno@gmail.com>