tmux: add systemd service
Add a systemd service so a tmux session is started at boot. It depends
on the graphical session to be fully initialized so that we inherit most
of the env variables as we would with a typical terminal emulator.

With this, a new directory was created and tmux.conf was moved
accordingly,

Signed-off-by: Nuno Sa <noname.nuno@gmail.com>