remote-connect: improve feedback on failure
In case freerdp fails to connect, we get no nothing (just nothing
happens). Try to improve it with a message and a dump of the logfile.
The reason why it's not easy to get feedback is because we 'exec'
freerdp and so we loose control over the process. The idea is now to
start the process in the background with "&" and then look for the
child status with 'tail --pidi $!'. We do this with a 5 seconds timeout
which means that if we timeout, then rdp is still running and we should
be fine (unless some strange thing happens and it blocks). If the child
dies before the 5 seconds, then it means we fail to connect and dump
the logs.

Note this 5 seconds are a first approach that seems to work reliable
(might need some adjustments in the future).

Signed-off-by: Nuno Sa <noname.nuno@gmail.com>