vscode: build: Improve build script
We can infer for the current .config which architecture we should
compile for and assume all defaults we were passing in through
tasks.json.

Moreover, the current changes will allow to greatly simplify the
tasks.json file as we do not need a task per platform but per
architecture (naturally we still proper defconfig tasks).

Signed-off-by: Nuno Sá <noname.nuno@gmail.com>