vscode: tasks: Simplify and unify tasks
With the recent changes in the build script

commit 1a477c3fdbd1 ("vscode: build: Improve build script"), we can now
remove a lot of tasks that can turn into generic ones (build, clean,
config..). We also introduced some new tasks so that the idea is now to
use this unique file across all the worktrees I have (instead of copy
paste and adapt in each tree).

Signed-off-by: Nuno Sá <noname.nuno@gmail.com>